split download scripts
